pyproject.toml
[project]
name = "mitigate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate==0.21.0",
    "datasets>=2.19.1",
    "diffusers==0.18.2",
    "fire>=0.7.0",
    "huggingface-hub==0.25.0",
    "matplotlib>=3.10.3",
    "numpy<2",
    "open-clip-torch>=2.32.0",
    "python-dotenv>=1.1.0",
    "scipy>=1.15.3",
    "torch==1.13.1",
    "torchaudio==0.13.1",
    "torchvision==0.14.1",
    "transformers==4.30.2",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu117" }]

[[tool.uv.index]]
name = "pytorch-cu117"
url = "https://download.pytorch.org/whl/cu117"
